Use retries for unmounting filesystem at the end.

+++ test/test_common.sh
@@ -6,14 +6,21 @@ function common_init()
 	mkdir mountpoint
 }
 
-function wait_for_fuse()
+function run_with_tries()
 {
-	local tries=100
-	while [ $tries -gt 0 ] && ! ls mountpoint >/dev/null 2>&1; do
+	local tries=$1
+	shift
+	while [ $tries -gt 0 ] && ! "$@"; do
 		(( --tries ))
 		sleep 0.01
 	done
-	if [ $tries -eq 0 ]; then
+
+	[ $tries -gt 0 ]
+}
+
+function wait_for_fuse()
+{
+	if ! run_with_tries 100 ls mountpoint >/dev/null 2>&1; then
 		echo "Failed to mount filesystem."
 		exit 1
 	fi
@@ -21,8 +28,10 @@ function wait_for_fuse()
 
 function common_cleanup()
 {
-	fusermount -u mountpoint
+	if ! run_with_tries 100 eval 'fusermount -u mountpoint && rm -rf mountpoint && [ ! -e mountpoint ]'; then
+		echo "Failed to remove mountpoint."
+		exit 1
+	fi
 	rm -rf rootdir
-	rm -rf mountpoint
 }
